Better (maybe) Linux Fix.
I tried a new approach to the Synaptics power state issue based on the understanding (to the best of my ability, not sure if I have this right) that one of the parameters of OnD0Exit is actually the target power state that the device is seeking to go to.

https://msdn.microsoft.com/en-us/library/windows/hardware/ff556803(v=vs.85).aspx

https://github.com/Microsoft/Windows-driver-samples/blob/master/sensors/SimpleDeviceOrientationSensor/device.cpp#L537

MrChromebox has tested this fix, and it seems to work, but I'm not sure if I've just made it so that the check fails every time. The thought was based on the list of power states here:

https://msdn.microsoft.com/en-us/library/windows/hardware/ff552421(v=vs.85).aspx

Hope being that on states 5, 6, or 7 (> 4), the Deep sleep would not be used, and this would allow one to reboot/shut down to Linux and have the touchpad work while still retaining Deep sleep in D3.